Move ResolveAddr to the protocol package
It's useful to clients and servers there, and reduces the dependencies
that trace/ would pull in when using the listen addr resolution
mechanism.
