+++ networking_test.go
@@ -11,30 +11,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/stripe/veneur/protocol"
 )
-
-func TestListenAddr(t *testing.T) {
-	tests := []struct {
-		input   string
-		network string
-		laddr   string
-	}{
-		{"udp://127.0.0.1:8200", "udp", "127.0.0.1:8200"},
-		{"tcp://:8200", "tcp", ":8200"},
-		{"tcp6://[::1]:8200", "tcp", "[::1]:8200"},
-		{"unix:///tmp/foo.sock", "unix", "/tmp/foo.sock"},
-		{"unixgram:///tmp/foo.sock", "unixgram", "/tmp/foo.sock"},
-		{"unixpacket:///tmp/foo.sock", "unixpacket", "/tmp/foo.sock"},
-	}
-	for _, test := range tests {
-		addr, err := ResolveAddr(test.input)
-		if !assert.NoError(t, err) {
-			continue
-		}
-		assert.Equal(t, test.network, addr.Network())
-		assert.Equal(t, test.laddr, addr.String(), "Address %#v not correct", addr)
-	}
-}
 
 func TestMultipleListeners(t *testing.T) {
 	srv := &Server{}
@@ -44,7 +22,7 @@ func TestMultipleListeners(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dir)
 
-	addrNet, err := ResolveAddr(fmt.Sprintf("unix://%s/socket", dir))
+	addrNet, err := protocol.ResolveAddr(fmt.Sprintf("unix://%s/socket", dir))
 	require.NoError(t, err)
 	addr, ok := addrNet.(*net.UnixAddr)
 	require.True(t, ok)
@@ -73,7 +51,7 @@ func TestConnectUNIX(t *testing.T) {
 	require.NoError(t, err)
 	defer os.RemoveAll(dir)
 
-	addrNet, err := ResolveAddr(fmt.Sprintf("unix://%s/socket", dir))
+	addrNet, err := protocol.ResolveAddr(fmt.Sprintf("unix://%s/socket", dir))
 	require.NoError(t, err)
 	addr, ok := addrNet.(*net.UnixAddr)
 	require.True(t, ok)
